requirements.txt
aiogram==3.13.1
rlottie-python==1.3.4
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
